pyproject.toml
[project]
name = "ring-dataset-gen"
version = "0.1.0"
description = "Reproducible computer-rendered dataset generator for ring inner-diameter estimation"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
qa = ["opencv-python>=4.5.0"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

